feat: restructure, update
- update from setup.py to pyproject.toml
- use pdm
- add support for "x" after the numbers in the decklists
