+++ mtgproxies/scryfall/__init__.py
@@ -1,0 +1,16 @@
+__all__ = [
+    "canonic_card_name",
+    "get_card",
+    "get_cards",
+    "get_faces",
+    "get_image",
+    "search",
+    "recommend_print",
+    "card_by_id",
+    "cards_by_oracle_id",
+    "oracle_ids_by_name",
+    "get_price",
+]
+
+from mtgproxies.scryfall.scryfall import oracle_ids_by_name, get_price, cards_by_oracle_id, card_by_id, recommend_print, \
+    search, get_image, get_faces, get_cards, get_card, canonic_card_name
