+++ pyproject.toml
@@ -1,3 +1,67 @@
+[project]
+dynamic = ["version"]
+name = "MTG-Proxy-Print-Assembler"
+dependencies = [
+    "numpy",
+    "matplotlib",
+    "requests",
+    "tqdm",
+    "fpdf2>=2.3.0",
+
+]
+requires-python = ">=3.9"
+classifiers = [  # Optional
+        # How mature is this project? Common values are
+        #   3 - Alpha
+        #   4 - Beta
+        #   5 - Production/Stable
+        "Development Status :: 4 - Beta",
+        "Topic :: Games/Entertainment :: Card Games",
+        # Pick your license as you wish
+        "License :: OSI Approved :: MIT License",
+        # Specify the Python versions you support here. In particular, ensure
+        # that you indicate whether you support Python 2, Python 3 or both.
+        # These classifiers are *not* checked by 'pip install'. See instead
+        # 'python_requires' below.
+        "Programming Language :: Python :: 3",
+]
+authors = [
+    {name = "Robin Kupper", email = "robin.kupper@rwth-aachen.de"},
+    {name = "Adam Bajger", email = "adambaj@seznam.cz"},
+]
+keywords = ["pdf", "proxies", "mtg", "magic-the-gathering", "mtg-cards", "scryfall", "decklist"]
+description="Print high-resolution MtG proxies."
+
+[project.urls]
+Homepage = "https://github.com/DiddiZ/mtg-proxies"
+
+[tool.setuptools.packages.find]
+include = ["mtgproxies", "scryfall"]
+
+
+
 [tool.black]
 line-length = 120
 target-version = ['py39', 'py310', 'py311', 'py312']
+
+[build-system]
+requires = ["setuptools >= 61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.flake8]
+ignore = []
+per-file-ignores = { "tests/*" = [] }
+
+max-line-length = 120
+docstring-convention = "google"
+
+[tool.flake8.known-modules]
+mtg-proxies=["mtgproxies","scryfall"]
+
+
+[isort]
+line_length = 120
+profile = "black"
+
+
+
